doc-intelligence/requirements.txt
streamlit==1.41.1
pymupdf==1.25.3
pytesseract==0.3.13
Pillow==10.4.0
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
openpyxl==3.1.5
xlrd==2.0.1
pandas==2.2.3
requests==2.32.3
tqdm==4.67.1
nltk==3.9.1
numpy==1.26.4
